+++ src/graphs/random_graph.rs
@@ -5,9 +5,9 @@ use super::digraph;
 
 pub fn get_random_graph(
     num_nodes: usize,
-    conn_prob: f32,
-    weight_mean: f32,
-    weight_sd: f32,
+    conn_prob: f64,
+    weight_mean: f64,
+    weight_sd: f64,
 ) -> digraph::Digraph {
     let mut rng = rand::thread_rng();
     let normal = Normal::new(weight_mean, weight_sd).unwrap();
@@ -16,7 +16,7 @@ pub fn get_random_graph(
 
     for i in 0..num_nodes {
         for j in 0..num_nodes {
-            if i != j && rng.gen::<f32>() > conn_prob {
+            if i != j && rng.gen::<f64>() > conn_prob {
                 let raw_weight = normal.sample(&mut rng);
                 let rounded_weight = (raw_weight * 100.0).round() / 100.0;
                 g.add_edge(i, j, rounded_weight);
